feat: change from api to async function
