+++ .gitignore
@@ -38,4 +38,3 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
-src/app/leaderboard/page.tsx
